chore: tentative factorization
