+++ client/plugin/hooks/use-market.ts
@@ -1,6 +1,7 @@
-import { useState, useEffect } from "preact/hooks";
 import { useContract } from "./use-contract";
 import { useTransactions } from "./use-mytransactions";
+import { encrypt } from "../helpers/poseidon";
+import { getSaleProof } from "client/plugin/helpers/snarks";
 
 // Notice functions are a property inside of useMarket() component
 // TODO: it should add to mytransactions history
@@ -45,9 +46,26 @@ export function useMarket () {
 		);
 	};
 
+	const sale = async (listingId, orderId, key, sharedKey, nonce, keyCommitment, sharedKeyCommitment) => {
+		const receiptId = encrypt(key, [ sharedKey.x, sharedKey.y ], nonce);
+		const saleProofArgs = {
+			receipt_id: receiptId,
+			nonce: nonce.toString(),
+			key_commitment: keyCommitment.toString(),
+			shared_key_commitment: sharedKeyCommitment.toString(),
+			shared_key: [ sharedKey.x, sharedKey.y ],
+			key: key
+		};
+		const saleProof = await getSaleProof(saleProofArgs);
+		const sale = await market.sale(...saleProof, listingId, orderId, {
+			gasLimit: 1000000,
+		});
+		console.log(`sale tx: `, sale);
+	}
 
 	return {
 		list,
-		delist
+		delist,
+		sale
 	};
 }
