+++ client/plugin/views/MyOrdersView.tsx
@@ -2,16 +2,16 @@ import { Transaction } from "ethers";
 import { FunctionalComponent, h } from "preact";
 import { useState } from "preact/hooks";
 import { Button } from "../components/Button";
-import { Listing, Order } from "../components/MyListingsContext";
+import {  } from "../components/MyListingsContext";
 import { getListingsWithOrdersFromAddress } from "../helpers/transactions";
 import { useContract } from "../hooks/use-contract";
 import { useListings } from "../hooks/use-listings";
 import { useSigner } from "../hooks/use-signer";
-import { OrderDetailsProps, ActiveSigner } from "../typings/typings";
-import { listingStyles, orderStyles } from "./MyListingsView";
+import { Listing, Order, OrderItemProps, ActiveSigner } from "../typings/typings";
+import { listingStyles, orderStyles } from "../helpers/theme";
 
 
-export const OrderDetails: FunctionalComponent<OrderDetailsProps> = (props: OrderDetailsProps) => {
+export const OrderItem: FunctionalComponent<OrderItemProps> = (props: OrderItemProps) => {
 	return (
 		<div style={orderStyles.order}>
 			{[
@@ -43,7 +43,7 @@ export function MyOrdersView () {
 				{myOrders.map((listing) => (
 					<div style={{ display: "grid", rowGap: "4px" }}>
 						{listing.orders.map((order) => (
-							<OrderDetails order={order} action={async () => await refund(listing.listing, order)} childrenAction={'refund'} buttonDisabled={!order.isActive} />
+							<OrderItem order={order} action={async () => await refund(listing.listing, order)} childrenAction={'refund'} buttonDisabled={!order.isActive} />
 						)
 						)}
 					</div>
